+++ src/plone/restapi/tests/test_serializer_converters.py
@@ -145,16 +145,16 @@ class TestJsonCompatibleConverters(TestCase):
 
     def test_python_datetime(self):
         value = DateTime('2015/11/23 19:45:55.649027 GMT+3').asdatetime()
-        self.assertEquals(u'2015-11-23T19:45:55+03:00',
+        self.assertEquals(u'2015-11-23T16:45:55+00:00',
                           json_compatible(value))
-        self.assertEquals('"2015-11-23T19:45:55+03:00"',
+        self.assertEquals('"2015-11-23T16:45:55+00:00"',
                           json.dumps(json_compatible(value)))
 
     def test_zope_DateTime(self):
         value = DateTime('2015/11/23 19:45:55.649027 GMT+3')
-        self.assertEquals(u'2015-11-23T19:45:55+03:00',
+        self.assertEquals(u'2015-11-23T16:45:55+00:00',
                           json_compatible(value))
-        self.assertEquals('"2015-11-23T19:45:55+03:00"',
+        self.assertEquals('"2015-11-23T16:45:55+00:00"',
                           json.dumps(json_compatible(value)))
 
     def test_date(self):