+++ src/plone/restapi/tests/test_services_navigation.py
@@ -29,13 +29,33 @@ class TestServicesNavigation(unittest.TestCase):
             self.portal, u'Folder',
             id=u'folder',
             title=u'Some Folder')
+        self.folder2 = createContentInContainer(
+            self.portal, u'Folder',
+            id=u'folder2',
+            title=u'Some Folder 2')
+        self.subfolder1 = createContentInContainer(
+            self.folder, u'Folder',
+            id=u'subfolder1',
+            title=u'SubFolder 1')
+        self.subfolder2 = createContentInContainer(
+            self.folder, u'Folder',
+            id=u'subfolder2',
+            title=u'SubFolder 2')
+        self.thirdlevelfolder = createContentInContainer(
+            self.subfolder1, u'Folder',
+            id=u'thirdlevelfolder',
+            title=u'Third Level Folder')
+        self.fourthlevelfolder = createContentInContainer(
+            self.thirdlevelfolder, u'Folder',
+            id=u'fourthlevelfolder',
+            title=u'Fourth Level Folder')
         createContentInContainer(
             self.folder, u'Document',
             id=u'doc1',
             title=u'A document')
         transaction.commit()
 
-    def test_navigation(self):
+    def test_navigation_with_no_params_gets_only_top_level(self):
         response = self.api_session.get('/folder/@navigation')
 
         self.assertEqual(response.status_code, 200)
@@ -46,12 +66,60 @@ class TestServicesNavigation(unittest.TestCase):
                 'items': [
                     {
                         u'title': u'Home',
-                        u'@id': u'http://localhost:55001/plone'
+                        u'@id': u'http://localhost:55001/plone',
+                        u'description': u'',
                     },
                     {
                         u'title': u'Some Folder',
-                        u'@id': u'http://localhost:55001/plone/folder'
+                        u'@id': u'http://localhost:55001/plone/folder',
+                        u'description': u'',
+                    },
+                    {
+                        u'@id': u'http://localhost:55001/plone/folder2',
+                        u'description': u'',
+                        u'title': u'Some Folder 2'
                     }
                 ]
             }
         )
+
+    def test_navigation_service(self):
+        response = self.api_session.get(
+            '/folder/@navigation',
+            params={
+                "expand.navigation.depth": 2
+            })
+
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(response.json()['items']), 3)
+        self.assertEqual(
+            response.json()['items'][1]['title'], u'Some Folder')
+        self.assertEqual(len(response.json()['items'][1]['items']), 3)
+        self.assertEqual(len(response.json()['items'][2]['items']), 0)
+
+        response = self.api_session.get(
+            '/folder/@navigation',
+            params={
+                "expand.navigation.depth": 3
+            })
+
+        self.assertEqual(
+            len(response.json()['items'][1]['items'][0]['items']), 1)
+        self.assertEqual(
+            response.json()['items'][1]['items'][0]['items'][0]['title'],
+            u'Third Level Folder')
+        self.assertNotIn(
+            'items', response.json()['items'][1]['items'][0]['items'][0])
+
+        response = self.api_session.get(
+            '/folder/@navigation',
+            params={
+                "expand.navigation.depth": 4
+            })
+
+        self.assertEqual(
+            len(response.json()['items'][1]['items'][0]['items'][0]['items']),
+            1)
+        self.assertEqual(
+            response.json()['items'][1]['items'][0]['items'][0]['items'][0]['title'], # noqa
+            u'Fourth Level Folder')